xml/parsing/manakai/peref-5.dat

#data
<!DOCTYPE a [
  %ho+ge;
]><a/>
#errors
1;1;no XML decl;;;s
2;3;entity not declared;;%ho+ge;;m
2;3;stop processing;;;i
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a [
  %ho+ge;
]><a/>
#checker ForValidator
#errors
1;1;no XML decl;;;s
2;3;entity not declared;;%ho+ge;;m
2;3;xml:not name;;ho+ge;m
2;3;stop processing;;;i
#document
| <!DOCTYPE a>
| <{} a>
